Converted to float for position changes on axis.

+++ impl/sdl2/inc/sdl2axis.hpp
@@ -12,7 +12,7 @@ class SDL2Axis : public Axis {
 public:
   SDL2Axis(SDL_Joystick* joystick, const uint32_t axisIndex);
   virtual ~SDL2Axis();
-  int16_t getPosition();
+  float getPosition();
   
 private:
   uint32_t mAxisIndex;
